Add files via upload
To-do:

Write up big preprocess function that takes regex, demotify, has a list of stop words, etc.

Do away with pandas.

Make it so the regression models are trained and run predictions on 1-5 stars, then stars are binned for the classification models
	Following that, the predictions from regression models are binned into -1, 0, and 1 like the classifiers.
	*Alternatively, we bin for regression too and we can just demonstrate later that we could have done a regression.

Optional: Switch to predict_proba on the models and use the CallibratedClassifierCV on models that don't have a predict_proba function. This will make the ensemble more accurate.